Automatically backed up by Learn

+++ db/migrate/002_create_costume_stores.rb
@@ -3,7 +3,7 @@ class CreateCostumeStores < ActiveRecord::Migration[5.1]
     create_table :costume_stores do |t|
       t.string :name
       t.string :location
-      t.integer :costume_inventory
+      t.float :costume_inventory
       t.integer :employee_count
       t.boolean :still_in_business
       t.datetime :start_time
